Made a TimeUtils class.
